+++ R/cirls.control.R
@@ -6,7 +6,7 @@
 #' @param maxit Integer giving the maximal number of CIRLS iterations.
 #' @param trace Logical indicating if output should be produced for each iteration.
 #' @param Cmat Constraint matrix specifying the linear constraints applied to coefficients. Can also be provided as a list of matrices for specific terms.
-#' @param lb,ub Lower and upper bound vectors for the linear constraints. Identical values in `lb` and `ub` identify equality constraints. Recycled if length is different than the number of constraints defined by `Cmat`.
+#' @param lb,ub Lower and upper bound vectors for the linear constraints. Identical values in `lb` and `ub` identify equality constraints. As for `Cmat` can be provided as a list of terms. If some terms are provided in `Cmat` but not in `lb` or `ub`, default values of 0 and Inf will be used, respectively.
 #' @param qp_solver The quadratic programming solver. One of `"quadprog"` (the default), `"osqp"` or `"coneproj"`.
 #' @param qp_pars List of parameters specific to the quadratic programming solver. See respective packages help.
 #'
@@ -28,29 +28,50 @@ cirls.control <- function (epsilon = 1e-08, maxit = 25, trace = FALSE,
     stop("value of 'epsilon' must be > 0")
   if (!is.numeric(maxit) || maxit <= 0)
     stop("maximum number of iterations must be > 0")
-  # Chech Cmat and prepare it
+
+  #----- Constraints preparation
+
+  # Chech Cmat has been provided
   if (is.null(Cmat)){
-    stop("Cmat must be provided")
-  } else {
-    # Get objects that cannot be passed through arguments
-    mt <- get("mt", envir = parent.frame(2))
-    x <- get("x", envir = parent.frame())
-    if (is.list(Cmat)){
-      Cmat <- clist2cmat(Cmat, mt, x)
-    } else {
-      if (ncol(Cmat) != ncol(x)){
-        stop("Cmat must have the same number of columns as the design matrix")
+    stop(paste0("Cmat must be provided. Note that 'control' and '...'",
+      " cannot be used at the same time, with 'control' being prioritised."))
+  }
+
+  # Get objects that cannot be passed through arguments
+  mt <- get("mt", envir = parent.frame(2))
+  x <- get("x", envir = parent.frame())
+
+  # If Cmat is a list, build matrix and bound vectors
+  if (is.list(Cmat)){
+
+    # Transform
+    Cmat <- clist2cmat(Cmat, mt, x)
+    ct <- attr(Cmat, "terms")
+
+    # Match bounds
+    bounds <- Map(function(b, def){
+      if (is.list(b)){
+        if (any(!names(b) %in% names(ct)))
+          warning("'ub' and 'lb' terms not found in 'Cmat' are dropped")
+        bvec <- rep(def, NROW(Cmat))
+        for (i in seq_along(b)) bvec[ct[[names(b)[i]]]] <- b[[i]]
+      } else {
+        bvec <- rep_len(b, nrow(Cmat))
       }
+      bvec
+    }, b = list(lb, ub), def = c(0, Inf))
+    lb <- bounds[[1]]
+    ub <- bounds[[2]]
+  } else {
+    if (ncol(Cmat) != ncol(x)){
+      stop("Cmat must have the same number of columns as the design matrix")
     }
-  }
-  # Check lb and recycle if needed
-  if (NROW(lb) != nrow(Cmat)){
+    if (is.list(ub) || is.list(lb)) stop(paste0("when 'Cmat' ",
+      "is provided as a matrix, 'lb' and 'ub' cannot be lists"))
     lb <- rep_len(lb, nrow(Cmat))
-  }
-  # Same with ub and recycle if needed
-  if (NROW(ub) != nrow(Cmat)){
     ub <- rep_len(ub, nrow(Cmat))
   }
+
   # Check that bounds are well specified
   if (any(lb > ub)){
     warning("lb should be lower than (or equal) ub")
@@ -59,7 +80,8 @@ cirls.control <- function (epsilon = 1e-08, maxit = 25, trace = FALSE,
     lb <- lb2
     ub <- ub2
   }
-  # Check irreducibility
+
+  # Check irreducibility of Cmat
   if (nrow(Cmat) > 1){
     chkc <- checkCmat(Cmat)
     if (length(chkc$redundant) > 0){
@@ -75,9 +97,13 @@ cirls.control <- function (epsilon = 1e-08, maxit = 25, trace = FALSE,
         "Consider using lb and ub to set equality constraints instead."))
     }
   }
+
+  #----- Other parameters
+
   # Prepare QP solver
   qp_solver <- match.arg(qp_solver, c("quadprog", "osqp", "coneproj"))
   qp_pars <- do.call(sprintf("%s.def", qp_solver), qp_pars)
+
   # Return
   list(epsilon = epsilon, maxit = maxit, trace = trace, Cmat = Cmat,
     lb = lb, ub = ub, qp_solver = qp_solver, qp_pars = qp_pars)