+++ R/shapeConstr.bs.R
@@ -5,7 +5,6 @@
 #
 ################################################################################
 
-#' @rdname shapeConstr
 #' @export
 shapeConstr.bs <- function(x, shape, ...){