+++ R/shapeConstr.ps.R
@@ -5,7 +5,6 @@
 #
 ################################################################################
 
-#' @rdname shapeConstr
 #' @export
 shapeConstr.ps <- function(x, shape, ...){